Return name in addition to URLs for related objects
Added name property to URLs so that clients don’t have to make an
additional call to get the name of an object. Previously, objects were
returned as a list of URLs:

  "films": [
      "http://swapi.co/api/films/1/",
      "http://swapi.co/api/films/2/"
  ]

Now objects are returned with name and url properties:

    "films": [
        {
            "url": "http://swapi.co/api/films/1/",
            "name": "A New Hope"
        },
        {
            "url": "http://swapi.co/api/films/2/",
            "name": "The Empire Strikes Back"
        }
    ]

Also, removed explicit definitions for some properties in the
serializers since they were already included in the Meta classes.

Fixes #9 and #29